P4ADEV-3775-taxonomyCode-validation-for-DebtPositionType added interface CategoryValidatorService
